+++ lib/src/draw_shape_options.dart
@@ -6,6 +6,7 @@ class DrawShapeOptions {
   const DrawShapeOptions({
     this.enable = false,
     this.onCompleteShape,
+    this.onClearShapes,
     this.newShapeColor =
         const Color(0xAA15781b), // default to lichess.org green
   });
@@ -15,6 +16,9 @@ class DrawShapeOptions {
 
   /// A callback for when shape drawing gesture is completed.
   final void Function(Shape shape)? onCompleteShape;
+
+  /// A callback for when the user clears all shapes.
+  final void Function()? onClearShapes;
 
   /// The color of the shape being drawn.
   final Color newShapeColor;